+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v2
 - [ ] add related articles, because I have multiple articles of "Me discovering Ruby ..." for example
+  - [ ] find related articles automatically by name
 - [ ] tags in the sidebar seem unsorted, introduce "Top tags" and "All tags" where they appear sorted, i would say
 - [ ] rebuild when a content file changes (not just when code changes doh)
 - [ ] add follow button for twitter